// cafeteria-frontend/src/models/SystemDefaults.ts
export default interface SystemDefaults {
  id: number;
  orderStartPeriodCount: number;
  orderStartPeriodType: number;
  orderStartRelativeTo: number;
  orderStartTime: string;
  orderEndPeriodCount: number;
  orderEndPeriodType: number;
  orderEndRelativeTo: number;
  orderEndTime: string;
  mealPrice: number;
  drinkOnlyPrice: number;
  squareAppId: string;
  squareLocationId: string;
}